Add support for clearing notifications
As per [1], tags and a notification message of "clear_notification" can
be used to clear a previous notification, which is useful when a
notification is about some warning state that has itself cleared. It
seems that notify-send can't actually clear notifications, but
fortunately we can use a simple call to gdbus instead. Handles clearing
multiple notifications with the same tag.

Tested with KDE Plasma 5.27.11.

[1] https://companion.home-assistant.io/docs/notifications/notifications-basic/#clearing